# app/openapi.yaml
openapi: 3.1.0
info:
  title: CB Service API
  description: API для CRUD по курсу CNY→RUB и фонового парсинга, с уведомлениями
    через WebSocket/NATS.
  version: 1.0.0
paths:
  /tasks/run:
    post:
      tags:
      - generator
      summary: Run Rate Generator
      operationId: run_rate_generator_tasks_run_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
  /items/:
    get:
      tags:
      - items
      summary: Read Items
      operationId: read_items_items__get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Item'
                type: array
                title: Response Read Items Items  Get
    post:
      tags:
      - items
      summary: Create Item
      operationId: create_item_items__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemCreate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /items/{item_id}:
    get:
      tags:
      - items
      summary: Read Item
      operationId: read_item_items__item_id__get
      parameters:
      - name: item_id
        in: path
        required: true
        schema:
          type: integer
          title: Item Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - items
      summary: Update Item
      operationId: update_item_items__item_id__patch
      parameters:
      - name: item_id
        in: path
        required: true
        schema:
          type: integer
          title: Item Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - items
      summary: Delete Item
      operationId: delete_item_items__item_id__delete
      parameters:
      - name: item_id
        in: path
        required: true
        schema:
          type: integer
          title: Item Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    Item:
      properties:
        date:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Date
        rate:
          type: number
          title: Rate
        id:
          type: integer
          title: Id
      type: object
      required:
      - rate
      - id
      title: Item
    ItemCreate:
      properties:
        date:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Date
        rate:
          type: number
          title: Rate
      type: object
      required:
      - rate
      title: ItemCreate
    ItemUpdate:
      properties:
        date:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Date
        rate:
          anyOf:
          - type: number
          - type: 'null'
          title: Rate
      type: object
      title: ItemUpdate
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError